added missing UART2 defs and cleanup
